requirements.txt
tgcrypto==1.2.3
Pyrogram==2.0.22
tortoise-orm